move App related things to APP folder
